Adds a striped map implementation
Includes an implementation of a striped map, where we can specify how
many different stripes (maps) to include, where we can lock on the
individual maps.  Rather than lock on the entire map, this way we can
have items partitioned into a different stripe based on a hash of the
key being used. This is intended to reduce the number of lock waits when
requesting resource info for specific nodes.