Only generate a self-signed cert, not a whole chain and keys
The user will have either supplied a private key or we can reuse their
client key as key material. In both cases, we don't actually need to
persist the self-signed certificate – we can just generate a fresh one
when the server starts as the key fingerprint will continue to match.