Ensure compute node does not open jobstore
During startup of a compute only node (using serve) the requester config
is retrieved, and as part of that a reference to the jobstore is
created.  This means that if we run a separate requester and compute,
the compute won't start because of the jobstore is already locked. The
bash-scripts avoid this by creating a new repo for the second node.

This PR is a short-term measure to only create the stores (compute and
execution) when explicitly asked to, and this is determined by the node
type when retrieving the config.

It is expected that this will be fixed in the node setup refactor.