Node registration when network=nats (#3535)
Introduces a requester endpoint to allow compute node to register their
existence at startup.

Compute nodes are responsible for supplying the mostly static
information about themselves at node startup time. This will include the
operating system, and the available resources etc (models.NodeInfo). The
dynamic data (such as currently used resources) will be sent separately
from this registration and so is not necessary during the initial
request.

Upon receiving the registration request the node info is added to the
persistent node store. This data is then served via the usual
nodeinfostore use-cases. Missing from the data will be:

* Current resource availability 
* Approval status 
* Any form of liveness info

These are to come in following PRs