Cancel authn input if the user presses Ctrl+C
Previously the system would wait for the user's input and not cancel
input if the user pressed Ctrl+C. This is because the signal was
intercepted and used to cancel the root context, but this context
isn't used by the terminal reading code and so had no effect.

The solution is to run the input in a separate goroutine and select on
the routine's output or the context cancellation.