Live compute-node resource tracking (#3556)
When using NATS, the compute node will now report latest resource
availability every 30 seconds to the requester node.  This info is then
added to any NodeInfo structures returned as part of requests to the
NodeInfoStore.

This is achieved by the NodeManager implementing both the Management API
and the NodeInfoStore API where it delegates to the persistent store and
adds the latest resource info. This resource info is held in a
Striped/Partitioned
map to reduce lock contention issues (e.g. we only have to lock the
bucket
where the key hashes to, not the whole map).

We might in later versions want to timestamp the resources (with a
server only timestamp) so that the server can be away how out of sync
the resource updates are.