Show node approval status in CLI
Shows the node approval status when running the `node list` command.
This will currently show as APPROVED for requester nodes, and PENDING
for compute nodes, but as these values are not currently used in
filtering this is okay.  Once we start filtering nodes based on their
approval status being APPROVED then we will need to decide on how we
auto-approve compute nodes.