Fix error on reading an empty tokens file
It's not clear to me why this file might end up empty – the test cases
try and catch all the possibilities but none result in an empty file,
but no behaviour seems to do this.

So anyway, in a defensive move that will hopefully catch any instances
of however this occurs, an empty tokens file is now treated as an
empty JSON map.