Validate BoltDB store schemas before use
This was the cause of the phantom "open: no such file or directory"
errors that we saw in #3481.

There is definitely still some issue with the config migration system
that causes this, but at least now we crash with a sensible error that
allows the problem to be pinpointed rather than an opaque error.