More gracefully handle invalid WASM specs
Previously an invalid WASM spec (with no EntryModule) would result in
a nil pointer in the input sources list. This now at least removes
that nil pointer.

But we still don't catch the actual problem (invalid spec) until we
are in the compute node, which is bad. Ideally we'd catch it much
earlier, or allow this method to more generically handle specs. I did
attempt a refactor to allow this, but having to plumb it into our gRPC
executor (unused) prevented this from working.