Store registration sentinel file alongside executions db
Currently the code is asking config for the repo directory to create a
file to denote successful registration.  This means that various places
(tests, devstack) then need to set this value when they provide their
own.

Instead, we will explicitly add a field to ComputeConfig and calculate
the value at runtime from the directory where the executions database is
stored.