Manually cleanup repo directory to work around flaky window test. (#3602)
On windows we've been seeing failing tests with directories already
existing, presumably from previous/parallel runs. This PR switches to
using a tempdir provided by the test framework so that it is cleaned up
when the test completes.

Edit: It appears this is a common problem on windows when removing
directories under certain conditions. We've not moved to manually
cleaning up the folder.