Fix job logs following to work with auth
Auth broke logs following because not only was there an incorrect cast
the setting of headers was not done correctly on the websocket request

To bring logs reading into the fold, we've had to refactor a little
bit and make Dial a first class method on the client. Unfortunately
there is a collision between the interface-typed Client and the
generic-typed Dialler, meaning that Dial now has to return just byte
slices and a utility is needed to parse these into strongly-typed
objects.