Clarifies error message when bacalhau encounters a newer repo version
When using an older version of bacalhau, in an environment with a repo
version that is higher than that supported by that version (e.g.
launching bacalhau 1.2.1 where the repo version has been upgrade to 3)
the error message is:

```
failed to initialize bacalhau repo at '/Users/ross/.bacalhau':
failed to check if repo exists: unknown repo version 3
```

The error message is now:

```
failed to initialize bacalhau repo at '/Users/ross/.bacalhau':
Unsupported repository version 3.
Bacalhau does not know how to read this configuration format and
you will need to upgrade to a newer version to upgrade this repository.
```