Approve/Reject node membership from CLI (#3594)
This PR adds the ability to approve/reject node membership to the
cluster, based on each node's approval state. The approve/reject
commands are similar and implemented using the same principle.

```
bacalhau node approve [nodeid] <-m message>

bacalhau node reject [nodeid] <-m message>
```

it is expected that this same approach can be used for `node pause` and
`node drain` etc.
Currently the system does nothing to react to the change of state, a
future PR will add pubsub for node events so that jobs can be scheduled
etc based on approval/rejection/join/leave events.

Additionally this PR resolves a small issue in AuthenticatingClient
which was proxying calls with the wrong methods, and adds a basenats
test suite for testing against a nats enabled devstack.