Allow nodes to call register on startup, even if previously registered
Currently we try and maintain state on the compute node about its
registration.  This can occassionally cause problems if there is any
disagreement with the node info kvstore, and we can't create them
transactionally.

This PR allows the compute node to register on each startup.  If nothing
else it will show an early indication of connectivity problems, and the
requester node knows only to complain for rejected nodes.