Add authentication to the Web UI.
This commit hooks up the Web UI to use our authentication APIs. It can
now query for a list of acceptable authentication methods and show the
user a web form for filling out the required credentials.

The UI will monitor requests and if it detects an Unauthorized status
will immediately open the auth flow. So as with the CLI, auth happens
automagically when required and not otherwise.