Adds a `node delete` command
Allows for nodes to be delete from the node info store. If the node is
found in the store, it is deleted, and then will not re-appear when the
requester node is re-started.  If however the compute node restarts, it
will re-register so this command is only useful if you know that nodes
will not be coming back.

In future if we want to block the node-id from ever being an entry, we
should add tombstone entries to the node info store that blog it's
registration.