Add error output to job selection exec hooks
Bacalhau compute nodes can be configured to call a script whenever
they receive a new job. If the script returns a non-zero error code,
the job is rejected.

Previously, the error message returned when one of these scripts
rejects a job is not helpful because it doesn't give much indication
of why the job was rejected.

Now, any text the script outputs onto STDERR to explain why it has
rejected the job will be passed back to the user in an error message.

This also updates the HTTP allowlist hook to do this.