Load WASM module dependencies if non-WASM input data is specified
There was previously a bug where if a user specified a job spec which
uses the WASM module dependency feature to load in WASM from multiple
modules, it would fail if there was more than a single dependent
module or if the WASM module to load is not specified first in the
input sources list.

This commit adds tests of this feature and ensures that WASM module
dependencies are loaded by examining the Alias of the InputSources.