Print ranking failures in the CLI
A bug in the BoltDB job store implementation meant that any
non-execution comments were not stored. This commit stores the error
from the scheduler in the job state as it used to be so that it can be
printed on the CLI.

It also only prints "Job Results by Node" if the job was actually
successfully placed on any node, and adds a test.