Node heartbeats (#3709)
Implements heartbeats for compute nodes, sending heartbeat messages to
the requester node over NATS PubSub. The server, upon receiving a
heartbeat updates the map of nodes to include the current server-side
timestamp.

Compute nodes using the heartbeat client, will continuously send
heartbeat messages every n seconds.

The heartbeat server receiving these heartbeats maintains a priority
queue, which dequeues oldest items (lowest timestamp) first. Every 5
seconds any item older than a specific timestamp is dequeued, and its
state either set to unhealthy (if it is the first missed heartbeat) or
unknown if it is the second. The default for timestamps is

* 30s since heartbeat - unhealthy
* 60s since heartbeat - unknown (node may be live but disconnected)

The next heartbeat sent by a unhealthy of unknown node will make it
healthy again and ready to receive work.

The current state of the node is added to the nodeinfo during a
Get/GetByPrefix/List call to the node info store. This means that the
liveness is dynamic and not persisted to the kvstore for node info.