Change the jobstore inprogress index to a composite key
Currently the inprogress index is just the ID of the job that is
inprogress.  This commit switches the inprogress index to a composite
key containing both the ID and the Type of the job.  This allows for
filtering of active jobs of a certain type without having to serde each
entire job referenced in the index.

We may at a future date want to introduce CompositeIndex type alongside
the current Index type.