Deprecates some IPFS flags (#3778)
In order to smooth the migration from an embedded IPFS node to a
self-hosted IPFS node, we will deprecate the flags used for the internal
node, rather than removing the functionality in one release.

This will allow for the functionality to be removed over two or more
minor version releases.

The Definition type has two new fields, a boolean Deprecated field and a
string DeprecatedMessage field. if the former is set then we mark the
flag as deprecated and show the second field.