Fixes a panic in the wasm executor when entrypoint is not known (#3779)
Currently we try to reference the `ExportedFunction` which is documented
as returning nil when the name is not found. Instead it returns a
non-nil object with an inner nil field which will cause a panic when
`Call`-ed.

This commit instead looks up the function before referencing it to
ensure it exists, returning an error early if not. This allows us to
remove the call to recover which was handling the panic() for us.

Fixes #3753