ExecutionTimeout to fail executions instead of jobs (#3974)
This work is a pre-requirement for job queues as the orchestrator is
currently failing jobs that take longer than ExecutionTimeout from the
time they were submitted. ExecutionTimeout should only fail executions
that have been running for too long, and not fail jobs. If a job haven't
started and has been in the queue for too long, then this timeout should
take effect as there will be other config to timeout of fail a job has
been in the queue for too long or has been retried too many times.

A side effect of this PR is that timed out executions will be retried by
the scheduler on other nodes until one succeed or no more nodes to
retry. This will be handled in next job queue related PRs

---------

Co-authored-by: Forrest <6546409+frrist@users.noreply.github.com>