refactor: define config instance. config is no longer global (#3959)
# Motivation
Previously we had a global config store. This is a bad pattern and
complicates many aspects of creating a node. We should instead have a
config instance that we pass to methods requiring config.

# Background

Bacalhau uses cobra to accept CLI flags and viper to manage
configuration state. Viper and Cobra may be used together by binding
viper to CLI flag definitions defined via cobra. Binding in this way
allows configuration to be specified using a mixture of defaults, config
files, environment variables, and flags. The priority viper treats these
different options is as follows with highest precedence first:

1. Flag
2. Env Var
3. Config File
4. Defaults

This mean that when no config file, env vars, or flags are provided, the
default settings are used. When a config file is provided, the values
present in the file override default values. When env vars are provided
they will override the defaults and/or the config file. When flags are
passed they will override any other medium providing configuration.

# Implementation

The first point of interaction with cobra and viper is in
`cmd/cli/root.go` where we access the global viper instance and bind it
to all persistent flags present on the root command. This configures the
viper instance with values related to the repo path, API, and logging.
From there, child commands (all other commands) may bind additional
flags to viper. For example, `cmd/cli/serve/serve.go` binds many
different flags that configure how the server operates. When a child
command is executed the following steps are taken to fully configure
viper and produce a final configuration state.

1. An instance of `config` is created with the global viper instance
2. The repo path set in `cmd/cli/root.go` is retrieved from viper
3. A repo is initialized or opened using the repo path and configuration
instance
1. When opening or initializing a repo, it first checks if a config file
is present under its path. If one is found it loads the values from the
file into the `config` instance.
2. Next the repo ensures paths to various files and directories are set
on the config instance. The **repo will modify the config to include
these paths**. These paths include:
      1. Authentication Tokens
      2. User private key
      3. executor plugin path
      4. compute storage (`executor_storages`) path
      5. update checker file path
      6. TLS certificate cache
      7. Libp2p private key
      8. orchestrator nats store
      9. orchestrator jobs database (`jobs.db`)
      10. compute store database (`executions.db`)
3. Once the repo has modified the config to contain the various path,
joined on its own path, it will ensure the paths are present on the
filesytem by creating them or stat-ing them.
4. The initialized repo and reconciled configuration file are then used
by the command to execute. The reconciled config
(`types.BacalhauConfig`) contains values specified by defaults, config
file, env vars, and/or flags merged in the order described above.

For the rest of the commands implementation services accept
`types.BacalhauConfig` or a subset of it to configure themselves. They
no longer depend on the previously global `config` package.

## Questions?

Various fields mentioned above contain paths to locations on disk where
values may be read/stored. At present these paths can be anywhere on the
filesystem - they need not live under the repo path. Since it is not
possible to define defaults for these fields ahead of time, we set them
based on the repo path when it is initialized.

In future iterations of this work, as described in the [Rethinking
Configuration](https://www.notion.so/expanso/Rethinking-Configuration-435fbe87419148b4bbc5119d413786eb#5e137f97d2f24078b0c13853bed0462c)
document, we intend to decouple the config from the repo. This can be
achieved by removing the ability to configure the paths mentioned above
from the config system and instead make them all relative to the repo
path with hard-coded values. In this way, the repo path can become a
field of the config that is provided to the repo when it initializes.

---------

Co-authored-by: frrist <forrest@expanso.io>