fully remove embedded ipfs (#4061)
This PR fully removes embedded IPFS from our code base, including tests
where we are now using `local` publisher and `localDirectory` storage
source instead.

## Testing
IPFS related tests are skipped unless IPFS connect endpoint is provided,
such as through `BACALHAU_NODE_IPFS_CONNECT` environment variable.

These tests are running in circleci test workflow for
liux-amd64-integration where we also run an IPFS daemon in that
environment and set IPFS connect configuration.

## Other Changes

Few other issues were surfaced when switching to local publisher that
had to be fixed, including:
- local publisher used a different compression logic than s3 publisher,
but both used the same downloader. The issue with the previous
compression is that it used absolute paths when setting file headers and
that resulted in failures with windows during decompression as it
resulted in invalid file names.
- repo clone is a weird storage source that copies repos from git to
ipfs, index to kv.bacalhau.org (no longer running), and tries to load
repos from ipfs if they are cached instead of git. It is a whole mess
and have been removed
- s3 integration tests didn't run in circelci due to a change in
determining if they should be skipped or not where we tried to write a
fake file to s3. The problem is circelci credentials only had read
permissions to a defined prefix. Changed the logic to try to read a
non-existing file, and run the test if we get 404 NoSuchKey, otherwise
we skip the test such as with 403 AccessDenied.