Optimize locking in producer and consumer NATS streaming client (#4068)
## What are we doing here ?

- With this PR, we are making sure we do locking correctly. 
- In addition, we come up with a good way to cancel the stream. We store
a `cancel` function as part of stream. This cancel function is attached
to the context which gets passed on to the engine, which serves the
stream (currently only the log stream). If a heartbeat response is
received from the consumer stating that it is no longer interested in
this stream, then we invoke this `cancel` function. There might be
scenarios where few logs will make it through after context cancelation
due to race conditions. This should be fine and won't result in `nil`
pointers or writing to a closed writer as we only close the writer after
source channel is closed.