mark timed-out executions as failed instead of cancelled (#4151)
For some reason, timed-out executions were marked as cancelled instead
of failed, which is wrong. Also this resulted in compute node calling
`OnCancelComplete` on the requester node, which is a noop. This means
the requester node will only mark the execution as failed when the
housekeeper kicks in which has a buffer of 2 minutes, instead of as soon
as the failure is reported by the compute node.

Previously, this job will be marked as failed after 2-2:30 minutes:
```
bacalhau docker --timeout 10 run ubuntu sleep 120
```

With this change it will marked as failed in ~10 seconds