Implement `-c/--config` flag (#4242)
Closes https://github.com/bacalhau-project/bacalhau/issues/4217 and
#4243

## Changes to Configuration Handling

The `config` package has been refactored to allow more flexible and
powerful configuration options. The configuration constructor now
accepts the following inputs:

1. **Map of Config Keys to Values:** Explicitly set configuration
values.
2. **Map of Config Keys to Flags:** Command-line flags mapped to
configuration keys.
3. **Map of Config Keys to Environment Variable Names:** Environment
variables mapped to configuration keys.
4. **Slice of Config File Paths:** Paths to configuration files, applied
in the order provided.
5. **Default Configuration Structure:** A fallback configuration
structure.

### Precedence of Configuration Sources

The precedence of these options, from highest to lowest, is as follows:
1. Map of config keys to values
2. Map of config keys to flags
3. Map of config keys to environment variable names
4. Slice of config file paths
5. Default configuration structure

This means that if a configuration key is defined in multiple sources,
the value from the source higher on this list will be used.

### Example Usage

Here’s an example of how to use the new configuration constructor:

```go
cfg, err := config.New(
	config.WithValues(configValues),
	config.WithFlags(configFlags),
	config.WithEnvironmentVariables(configEnvVar),
	config.WithPaths(configFiles...),
	config.WithDefault(config.ForEnvironment()),
)
```

A more detailed example specific to Bacalhau can be found
[here](https://github.com/bacalhau-project/bacalhau/pull/4242/files#diff-32943909f6f67bb12e2713415faaccb08dd368722df4c6cc4e690213d652a3d3R59).

### Handling of Configuration Files in Bacalhau
When using Bacalhau, configuration files are treated differently
depending on their location:

- **`--config` flag**: When multiple configuration files are provided
via the --config flag, Bacalhau merges the values from these files. The
order in which the files are listed determines the precedence, with
later files overriding earlier ones.

- **Files in `~/.bacalhau` and `~/.config/bacalhau`**: These files are
not merged, and only used a file(s) isn't specified by the `--config`
flag. . Instead, Bacalhau prioritizes `~/.config/bacalhau`. If a config
file is found in `~/.config/bacalhau`, it is used exclusively. If not,
Bacalhau falls back to `~/.bacalhau`. If neither file is present,
Bacalhau does not load a config file.

### Precedence with Flags and Configuration Files
- **Flags vs. Config Files**: Command-line flags always take precedence
over values specified in configuration files. For example, in the
command:
```bash
bacalhau -c config.yaml -c override.yaml --api-host=1.1.1.1 serve
```
The client host will be set to 1.1.1.1, even if this key is present in
config.yaml or override.yaml.

- **Multiple Flags**: If multiple flags and config files are provided,
the --config flag still takes precedence. For example:
```bash
bacalhau -c config.yaml -c override.yaml --api-host=1.1.1.1 -c node.clientapi.host=2.2.2.2 serve
```
Here, the client host will be set to 2.2.2.2 as specified by the -c
node.clientapi.host=2.2.2.2 flag.

## Changes to Agent API
In order to implement proper testing of the bacalhau configuration
package the `agent` API was extended with a new endpoint
`/api/v1/agent/config` which returns the current configuration used by
an agent as json. This API allows bacalhau to be started with a specific
configuration and then queried to make assertions on the configuration
being used. `test/config.sh` implements coverage of this API and the
config package.

## Changes to `config list`
Previously this command returns a list of config keys and their default
values. It now returns a list of config keys and their descriptions.

---------

Co-authored-by: frrist <forrest@expanso.io>