drat: first pass at decoupling the repo and config paths (#4330)
closes: https://github.com/bacalhau-project/bacalhau/issues/4219

## Background:
At present the bacalhau config contains several fields that allow paths
to location on disk to be set, these fields include, but are not limited
to:
- KeyPath:
https://github.com/bacalhau-project/bacalhau/blob/50831fa55c96660fdc573aab4a4441f10d9f6000/pkg/config/types/config.go#L14
- ExecutorPluginPath:
https://github.com/bacalhau-project/bacalhau/blob/50831fa55c96660fdc573aab4a4441f10d9f6000/pkg/config/types/node.go#L21
- ComputeStoragePath:
https://github.com/bacalhau-project/bacalhau/blob/50831fa55c96660fdc573aab4a4441f10d9f6000/pkg/config/types/node.go#L24
- StoreDir:
https://github.com/bacalhau-project/bacalhau/blob/50831fa55c96660fdc573aab4a4441f10d9f6000/pkg/config/types/node.go#L123

If a user doesn't set a value for these fields in the config, bacalhau
will set them to default values under the repo path:

https://github.com/bacalhau-project/bacalhau/blob/50831fa55c96660fdc573aab4a4441f10d9f6000/pkg/repo/fs.go#L216-L229

This setup tightly couples the repo and config: in order for a config to
be correctly populate it must first be modified by the repo to set paths
accordingly in the event that are not provided. In order for a repo to
be initialized it requires a config.
This means the vanilla config returned by config.New() isn't actually
usable by a bacalhau node until it has been passed to a repo.


### Proposed Change
This draft removes usage of `*Path` fields defined in the bacalhau
configuration structure. Meaning the config file no longer dictates the
location on disk of various files and directories. Instead these paths
are defined as constants declared with in the config package. When a
service such as transport, compute, or requester require a path on disk
they may retrieve it from the config package, at which point the value
will be created, or returned if it already exists.

Exceptions:
The following `*Path` fields will still remain present in the config:
- All paths in the TLSConfiguration, these values are optional when
running bacalhau.
- `Metrics.EventTracerPath` - we currently always set this field to
`os.DevNull` - perhaps it would make sense to remove it entirely and its
related event tracer since its never used.
- `AuthConfig.AccessPolicy` - bacalhau provides a verity of access
policies in the form of .rego files, and users will likely want to
fiddle with this value, further its optional.
- `AuthentiatorConfig.PolicyPath` - same story as `AccessPolicy`

---------

Co-authored-by: frrist <forrest@expanso.io>