Introduce APIError for Sync Calls from CLI to Orchestrator Node (#4366)
This PR aims at introducing APIError, which is responsible for
encapsulating all Synchronous error for requestes from CLI to
Orchestrator Node.

We introduce a CustomHttpErrorHandler, which is responsible for
translating errors to correct response status code. For example if a job
does not exist, we should return 404 and not 500.

in addition, to this we also make sure that CLI outputs are clean
outputs and not involving the HTTP Response codes. There is not value in
exposing HTTP Status Codes on CLI as it adds no values.


closes #4234