fix panic due to nil execution store (#4501)
The problem is we we don't fail the node properly when we fail to open
execution store's boltdb, which can happen if we already have a node
running using the same repo. This can be easily reproduced by running
two nodes on separate ports but using the same rep, or just trying to
run a node while devstack is running:

```
bacalhau serve --compute --orchestrator --api-port 1122

bacalhau serve --compute --orchestrator --api-port 2233
```

Now we get a clearer error of why we failed to start. It is not the best
error as it exposes implementation details and doesn't tell exactly what
is happning and how to fix things, but at better than a panic at least
and time is tight to improve the propagation of this error for now

```
failed to configure compute node: failed to create execution store: timed out while opening database, file "/Users/walid/.bacalhau/compute/state_boltdb.db" might be in use
```

Closes #4491