introduce system.GlobalConfig + migrate installation id (#4498)
After another look, we actually do persist the installationID in the
`config.yaml` file in v1.4.0. The bug was that if no installationID
existed, we kept on generating new ones with each run and not persisting
it. So if `config.yaml` has an installation, we should migrate it to
`~/.config/bacalhau/installation_id` which is what this PR does.

Few changes were required to achieve _and_ test this:
1. Moved `GetInstallationID()` from `config` to `system` package as the
installationID is not really configurable
2. Faced import cycle that forced me to move `system/tracer.go` to
`telemetry/tracer.go`, which is what is causing a lot of files to change
in the PR
3. Introduced `system.GlobalConfig` interface to allow mocking the
`ConfigDir()` and be able to test the migrations properly as we don't
want the tests to actually update `~/.config/bacalhau`