Slightly improve API error handling
This commit tries to fix the "FromHttpResponse" function. Returning 2 errors from
a function can be confusing and uncommon in Golang, especially if the function
only returns these 2 errors.

The new code code will return one error, which is always an APIError, and
it also improves the error messaging wording, adding more details in the output.

Several more tweaks can be done to the current error message handling
if needed.