fix orchestrator showing as disconnected (#4567)
orchestrator node doesn't heartbeat which results in its state showing
as disconnected in the webUI and http APIs. This PR fixes this by having
the orchestrator node treats itself as connected regardless of
heartbeats or not. We might have to introduce orchestrator level
heartbeats in the future when we introduce multiple orchestrators

### Testing done
- Unit tests
- Verified WebUI is showing the right state
- Verified compute nodes are still marked as disconnected when they go
down